+++ core/designsystem/src/main/kotlin/dev/hossain/devicecatalog/core/designsystem/component/DeviceCatalogTopAppBar.kt
@@ -77,7 +77,7 @@ fun DeviceCatalogTopAppBar(
 private fun DeviceCatalogTopAppBarPreview() {
     DeviceCatalogAppTheme {
         DeviceCatalogTopAppBar(
-            title = "Device Catalog",
+            title = "Android Catalog",
         )
     }
 }